add: support for microsoft v1 endpoints which requires resource for token aud (#421)
* add: support for microsoft which requires resource for token aud

Resource parameter depicts the identifier of the WebAPI that your client wants to access on behalf of the user. Most flows in OAuth involve 4 parties, the resource owner (aka user), the client (aka app), the authority (aka identity provider) and the resource (aka webapi). The audience of the access token that the authority generates is the resource identifier.

This allows us to authenticated with microsoft v1 endpoints

* fix: only add resource to retrieveType form

* fix: change from resource to scope

* fix: remove dead code

* Revert "fix: remove dead code"

This reverts commit 7398b1dc7db4b153467649148ec91cdec4745770.

* fix: misconfigured header